requirements.txt
streamlit
pandas
plotly
openpyxl
Pillow
kaleido

